requirements.txt
nova-act
python_dotenv